requirements.txt
flask
tensorflow
numpy
opencv-python
scikit-learn
tqdm
joblib
matplotlib
seaborn
pillow
gunicorn
